+++ jax_cem/datastructures/__init__.py
@@ -1,1 +1,2 @@
+from .trails import *  # noqa F403
 from .structures import *  # noqa F403